refactor: reset package structure to read more easily
