+++ src/main/java/tosiltosil/backend/common/domain/CustomException.java
@@ -1,4 +1,4 @@
-package tosiltosil.backend.common.exception;
+package tosiltosil.backend.common.domain;
 
 public abstract class CustomException extends RuntimeException {
