+++ src/main/java/tosiltosil/backend/common/logging/filter/TracingFilter.java
@@ -1,4 +1,4 @@
-package tosiltosil.backend.common.logging;
+package tosiltosil.backend.common.logging.filter;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
