+++ src/main/java/tosiltosil/backend/common/web/handler/ErrorResponse.java
@@ -1,4 +1,4 @@
-package tosiltosil.backend.common.exception;
+package tosiltosil.backend.common.web.handler;
 
 public record ErrorResponse(
         String customCode,
