+++ src/main/java/tosiltosil/backend/module/HealthCheckApiSpecification.java
@@ -1,4 +1,4 @@
-package tosiltosil.backend.common.monitoring;
+package tosiltosil.backend.module;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
