+++ src/main/java/tosiltosil/backend/module/HealthCheckController.java
@@ -1,4 +1,4 @@
-package tosiltosil.backend.common.monitoring;
+package tosiltosil.backend.module;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
